+++ test/End_to_end_test/bearer_test.js
@@ -29,6 +29,7 @@
 
 var chromedriver = require('./driver');
 var service = chromedriver.get_service();
+var error_handler = chromedriver.error_handler;
 var webdriver = chromedriver.webdriver;
 var By = webdriver.By;
 var until = webdriver.until;
@@ -37,7 +38,7 @@ var chai = require('chai');
 var expect = chai.expect;
 
 const TEST_TIMEOUT = 1000000; // 1000 seconds
-const LOGIN_WAITING_TIME = 3000; // 3 second
+const LOGIN_WAITING_TIME = 1000; // 1 second
 
 /******************************************************************************
  *  configurations needed
@@ -163,28 +164,36 @@ var get_token_for_resource = (resource, done) => {
     // we only need to enter the user name and password if we haven't logged in yet
     if (!client_already_logged_in) {
       driver.wait(until.titleIs('Sign in to your account'), 10000); 
-      var usernamebox = driver.findElement(By.name('login'));
+      var usernamebox = driver.findElement(By.name('loginfmt'));
       usernamebox.sendKeys(test_parameters.username);
+      usernamebox.sendKeys(webdriver.Key.ENTER);
       var passwordbox = driver.findElement(By.name('passwd'));
       passwordbox.sendKeys(test_parameters.password);
       driver.sleep(LOGIN_WAITING_TIME);
+      passwordbox = driver.findElement(By.name('passwd'));
       passwordbox.sendKeys(webdriver.Key.ENTER);
       client_already_logged_in = true;
     }
   }).then(() => {
-    done();
+    driver.wait(until.titleIs('Sign in to your account'), 5000).then(
+      ()=>{ driver.findElement(By.id('idBtn_Back')).click().then(()=>{done();}); },
+      ()=>{ done();}
+    );
   });
 };
 
 var checkResult = (test_app_config, result, done) => {
   var server = require('./app/api')(test_app_config);
-
   driver.get('http://localhost:3000/callApi')
   .then(() => {
-    driver.wait(until.titleIs('result'), 10000);
+    driver.wait(until.titleIs('result'), 10000).catch((ex) => {
+      error_handler(ex, server, done);
+    });
     driver.findElement(By.id('status')).getText().then((text) => { 
       expect(text).to.equal(result);
       server.shutdown(done);
+    }).catch((ex) => {
+      error_handler(ex, server, done);
     });
   });
 };